test: Add functional and unit tests
