+++ backend/parsers/__init__.py
@@ -1,0 +1,2 @@
+# Parsers package
+
