+++ backend/requirements.txt
@@ -2,3 +2,6 @@ Flask==3.0.0
 Flask-CORS==4.0.0
 notion-client==2.2.1
 python-dotenv==1.0.0
+pytest==7.4.3
+pytest-cov==4.1.0
+pytest-mock==3.12.0
