+++ backend/tests/functional/__init__.py
@@ -1,0 +1,2 @@
+# Functional tests package
+
