+++ backend/tests/unit/__init__.py
@@ -1,0 +1,2 @@
+# Unit tests package
+
