+++ backend/utils/__init__.py
@@ -1,0 +1,2 @@
+# Utils package
+
