+++ frontend/package.json
@@ -8,6 +8,11 @@
     "react-scripts": "5.0.1",
     "axios": "^1.6.0"
   },
+  "devDependencies": {
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/jest-dom": "^6.1.0",
+    "@testing-library/user-event": "^14.5.1"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
