+++ frontend/src/__tests__/README.md
@@ -1,0 +1,28 @@
+# Tests Frontend
+
+## Structure
+
+- `components/` : Tests des composants React
+- `hooks/` : Tests des hooks personnalisés
+
+## Exécution
+
+```bash
+# Tous les tests
+npm test
+
+# En mode watch
+npm test -- --watch
+
+# Avec couverture
+npm test -- --coverage
+```
+
+## Tests disponibles
+
+### Tests de composants
+- `ChatPage.test.js` : Tests du composant principal de chat
+
+### Tests de hooks
+- `useChatForm.test.js` : Tests du hook de formulaire
+
